basic work to refactor connect classes. probably doesn't pass all tests
